fix: use internal classes
