+++ tests/TestCase/Controller/Admin/AdministrationBaseControllerTest.php
@@ -1,9 +1,8 @@
 <?php
 namespace App\Test\TestCase\Controller\Admin;
 
+use App\Controller\Admin\AdministrationBaseController;
 use App\Controller\Admin\RolesController;
-use App\Test\Utils\AdminBaseController;
-use App\Test\Utils\WrongAdminBaseController;
 use BEdita\WebTools\ApiClientProvider;
 use Cake\Http\Exception\UnauthorizedException;
 use Cake\Http\Response;
@@ -17,18 +16,8 @@ use Cake\TestSuite\TestCase;
  */
 class AdministrationBaseControllerTest extends TestCase
 {
-    /**
-     * Test subject
-     *
-     * @var \App\Test\TestCase\Controller\Admin\AdminBaseController
-     */
     public $AdministrationBaseController;
 
-    /**
-     * Test subject
-     *
-     * @var \App\Controller\Admin\RolesController
-     */
     public $RlsController;
 
     /**
@@ -61,7 +50,10 @@ class AdministrationBaseControllerTest extends TestCase
 
         $config = array_merge($this->defaultRequestConfig, []);
         $request = new ServerRequest($config);
-        $this->AdministrationBaseController = new AdminBaseController($request);
+        $this->AdministrationBaseController = new class ($request) extends AdministrationBaseController
+        {
+            protected $resourceType = 'applications';
+        };
         $this->client = ApiClientProvider::getApiClient();
         $adminUser = getenv('BEDITA_ADMIN_USR');
         $adminPassword = getenv('BEDITA_ADMIN_PWD');
@@ -79,7 +71,11 @@ class AdministrationBaseControllerTest extends TestCase
     {
         $config = array_merge($this->defaultRequestConfig, $cfg);
         $request = new ServerRequest($config);
-        $this->RlsController = new RolesController($request);
+        $this->RlsController = new class($request) extends RolesController
+        {
+            protected $resourceType = 'roles';
+            protected $properties = ['name'];
+        };
         $this->client = ApiClientProvider::getApiClient();
         $adminUser = getenv('BEDITA_ADMIN_USR');
         $adminPassword = getenv('BEDITA_ADMIN_PWD');
@@ -191,7 +187,11 @@ class AdministrationBaseControllerTest extends TestCase
 
         $config = array_merge($this->defaultRequestConfig, []);
         $request = new ServerRequest($config);
-        $this->AdministrationBaseController = new WrongAdminBaseController($request);
+        $this->AdministrationBaseController = new class ($request) extends AdministrationBaseController
+        {
+            protected $resourceType = 'wrongtype';
+        };
+
         $this->AdministrationBaseController->index();
         $viewVars = (array)$this->AdministrationBaseController->viewVars;
         foreach ($keys as $expectedKey) {
