+++ tests/TestCase/Controller/Admin/EndpointsControllerTest.php
@@ -1,7 +1,7 @@
 <?php
 namespace App\Test\TestCase\Controller\Admin;
 
-use App\Test\Utils\EndsController;
+use App\Controller\Admin\EndpointsController;
 use BEdita\WebTools\ApiClientProvider;
 use Cake\Http\ServerRequest;
 use Cake\TestSuite\TestCase;
@@ -13,11 +13,6 @@ use Cake\TestSuite\TestCase;
  */
 class EndpointsControllerTest extends TestCase
 {
-    /**
-     * Test subject
-     *
-     * @var \App\Test\TestCase\Controller\Admin\EndsController
-     */
     public $EndsController;
 
     /**
@@ -50,7 +45,11 @@ class EndpointsControllerTest extends TestCase
 
         $config = array_merge($this->defaultRequestConfig, []);
         $request = new ServerRequest($config);
-        $this->EndsController = new EndsController($request);
+        $this->EndsController = new class($request) extends EndpointsController
+        {
+            protected $resourceType = 'endpoints';
+            protected $properties = ['name'];
+        };
         $this->client = ApiClientProvider::getApiClient();
         $adminUser = getenv('BEDITA_ADMIN_USR');
         $adminPassword = getenv('BEDITA_ADMIN_PWD');
