--- tests/Utils/UserProfileControllerSample.php
@@ -1,11 +1,0 @@
-<?php
-namespace App\Test\Utils;
-
-use App\Controller\UserProfileController;
-
-/**
- * @uses \App\Controller\UserProfileController
- */
-class UserProfileControllerSample extends UserProfileController
-{
-}
